TensorFlow: normalise path before using (NFCI)
This normalises the path before using it in the
`add_target_swift_library`.  This is important for platforms where the
path separator is not the Unix path separator (i.e. `/`), such as
Windows.  CMake will not normalise the path when globbing which can
result in unusable paths being passed to the function:

  ```
  Syntax error in cmake code when parsing string

  S:\3\s/tensorflow-swift-apis/Sources/TensorFlow/Operators/Comparison.swift

  Invalid character escape '\3'.
  ```

This should enable the glob to function properly in such environments.