[CodeCompletion] Suggest local variables in a multi-clause guard statement
Previously, local optional binding values aren't suggested in subsequent
conditions in guard statement. That was because, if the condition
contains code-completion token, the parser stops parsing and add implicit
body with the end loc of the condition which is, in this case, the
position of the code-completion token. As a result, since code-completion
location is in the body range, namelookup::FindLocalVal refuses to emit
valuses in conditions.

This patch fixes the issue by ignoring implict body in `FindLocalVal`.

rdar://problem/28482216
(cherry picked from commit 90f7f34a35d3d4602d5e70b30ddb579ef5b2d014)