SILGen: Fix crash when a closure captures two values with dynamic 'Self' type
We give special treatment to a capture of a value with dynamic 'Self'
type, by adding it at the end of the capture list. This ensures that
IRGen can recover the 'Self' metadata from this parameter if needed.

However, the code would crash with an assertion if there were multiple
captured values having the dynamic 'Self' type. This was slightly more
difficult to spell before SE-0068, but it was still possible if you
tried; now its completely trivial.

Fixes <https://bugs.swift.org/browse/SR-11928>, <rdar://problem/57823886>.