IRGen: Correctly compute instanceStart for a Swift class that starts with an empty field and is followed by a resilient field
In this case we used to compute an instance start of zero.

enum ZeroSizedEnum {
  case one
}

public class A {
  var y = ZeroSizedEnum.one
  var d = ResilientThing()
}

We would compute an ivar and field offset of zero for y. When these values are
slid because of the zero instanceStart value we would crash because the
ivar offset and field offset disagreed and we would write to the ivar
offset which is stored in constant memory.

rdar://58458169