TensorFlow: remove the unnecessary code in CTensorFlow
This removes the local CTensorFlow library which had dead code and a
decent chunk of complexity to import the TensorFlow interfaces.  As a
simplifying step towards extracting the TensorFlow module, this
introduces a second copy of the tensorflow-swift-apis code for the
CTensorFlow module.

The user visible parts of this change amount to restructuring the
tensorflow headers and locations.  The headers are now installed to
`lib/swift/tensorflow` and replicate the system header location layout,
of `tensorflow/c/...`.  This allows for re-use of a system-wide
installation of TensorFlow on Linux or a standalone packaged library on
Darwin.

This sets the stage to move towards externalising the build for the
TensorFlow module by removing the dependency on a library out-side of
the directory.  The build rules are largely replicating the build
infrastructure which has been constructed in the swift-apis repository.