Remove warning for conflicting implementationOnly imports
This warning occurred in the case that you had multiple files in a
module, some that used `@_implementationOnly import Foo` and some that
used `import Foo`. The downside of having this warning is that users
cannot represent in which files `Foo` has to be imported publicly vs
which files are only using it in an implementation only capacity. The
other downside is that if users go to add a public dependency on `Foo`
to their module, they would have to change all the other
`@_implementationOnly import Foo` calls in the module to no longer have
the annotation.