[5.2] IRGen: Correctly compute instanceStart for a pure Swift class that starts with an empty field and is followed by a resilient field
Description:
We used to crash for classes that have an empty and a resilient field
during intialization if the object was in the shared cache.

class CrashInInit {
var empty = EmptyStruct()
var resilient = ResilientThing()
}

What happened was that for such a class we we would compute a
instanceStart of 0. The shared cache builder would then slide the value
of the constant ivar offset for the empty field from 0 to 16. However,
the field offset for empty fields is assumed to be zero and the runtime
does not compute a different value for the empty field and so the field
offset for the empty field remains 0. The runtime then trys to reconcile
the field offset (0) and the ivar offset (16) trying to write to the ivar
offset. However, the ivar offset is marked as constant and so we
crashed.
This can be avoided by correctly computing the instanceStart for such a
class to be 16 such that the shared cache builder does not update the
value of the empty field.

Scope: This bug has existed in swift 5.1.

Testing: Test added to regression suite

rdar://58458169