[partial-apply-combiner] Fix handling of indirect parameters.
I do not believe that this code was ever correct and we just got lucky since in
the vast majority of cases where this optimization kicks in we also eliminate
the underlying partial apply/allocations.

To make sure that we can properly test this code and make sure that we do not
run into this problem again, I added an option to SILCombinerApplyVisitors that
disables partial apply deletion so we can verify that we do not rely on
eliminating the underlying partial apply for correctness.

<rdar://problem/57855082>

----

To be more specific, partial_apply expects that it will take all of its input
arguments at +1. This includes parameters that are passed @in or @in_guaranteed
to the underlying function. Just to summarize the expected semantics are that:

1. If the underlying function takes the parameter @in then the partial_apply
forwarder allows the underlying function to consume the argument.

2. If the underlying function takes the parameter @in_guaranteed then the
partial_apply forwarder itself will destroy the underlying value.

In terms of the partial_apply combiner this means that unless we can guarantee
that the partial_apply is eliminated as a result of our transformation (which we
can not givne the way this utility is written), we must introduce new copies of
@in, @in_guaranteed partial applied arguments and lifetime extend those copies
to the new direct apply that we are creating. In the case of the parameter being
in_guaranteed we insert a destroy after that call site to ensure that our new
copy is cleaned up. If we have an @in parameter then we allow the underlying
function to clean up the value.