Correctly destroy the original outer object
As pointed out in review comments by Slava, the previous code reused the `src`
and `srcType` variables as it unwrapped boxes and then ultimately released
whatever those pointed to.  Now I keep the original values around as
`originalSrc` and `originalSrcType` and use those in the cleanup code.

Here's how the flags are handled in this function now:
	* TakeOnSuccess always destroys the outer original object using the new `originalSrc*` variables.
  * DestroyOnFailure is never passed into this code, so can be ignored.
	* Unconditional is never passed into this code, so can be ignored.