build-script-impl: reorganise the option
This reorders the option list.  It tries to categorise them by area and
then sort them alphabetically.  This should help work through the list
and remove the now defunct options as well as try to reduce the set of
options into different build configurations that we can support and try
to organise the CMake invocations properly.

There is still a group of uncategorised options that should get sorted
out at some point.  However, this is still a huge step forward towards
simplifying (and hopefully eradicating) this monstrous script.