[AutoDiff] Add JVPs for tgmath functions.
Add JVPs for tgmath functions except `remainder(_:_:)` and `fmod(_:_:)`.
Add tests.

JVP/VJP functions for unary tgmath functions are identical, but there is no way
to register a derivative function as both a JVP/VJP of the same original
function. TF-1111 tracks this; the issue will naturally resolve when linear
functions and transposition are done and VJPs are removed.

Partially resolves TF-1108.