build-script: generalise the darwin handling a bit
There were cases for darwin targets that differed only by platform.
Update the logic to be more generic.  We should be able to specify the
Darwin options on non-Darwin targets and not have problems.  As a
result, it should be possible to fold away the special case for Darwin
platforms here.