build: use the unified build for compiler-rt/libc++
Use the runtimes build infrastructure that was setup by @llvm-beanz to
build libc++ and compiler-rt as part of the LLVM build rather than
building them as standalone components.  This simplifies the dependency
management and the logic for building/installing.

In order to build the complete compiler-rt for the target, it is
necessary to enable the libc++ runtime for building XRay.