[CSRanking] Consider solutions with different kinds of overloads as incomparable
If choices are not of the same kind it means that we can't really
compare the solutions e.g.

```swift
func foo(_: Int) -> (count: Int) {}
func foo(_: String) -> [Int] {}

_ = foo(#^placeholder^#).count
```

Here (for the same locator) one choice would be an array `[Int]`,
and another one would be a tuple `(count: Int)`, both have `.count`
with is referenced completely differently - a property in case of
an array and `element with label 'count:'` in case of a tuple.

Resolves: rdar://problem/49712598