[pa-combiner] Disable handling of indirect parameters if combining an immediately escaping closure.
In this commit, I am disabling this code path for cherry-picking purposes. It
should not hit perf since in most cases when one invokes a partial_apply right
after creating it, one will have a non-escaping function. So to hit this
micompile one needs to create an escaping closure and invoke it once before
escaping it. That seems pretty unusual.

That being said, we should not miscompile this pattern.