[test] Adjust new Substring test
We don’t have an easy way to check for the runtime version of the stdlib, so a two-way check for behavioral changes isn’t feasible. (Checking for the OS version isn’t good enough.)

Only check for the new behavior, and only when we know for sure that it’s available.