Implement Swift serialization and deserialization of Clang types.
As part of this, refuse to import Clang function-pointer types
as `@convention(c)` if we can't serialize them.  We don't expect
this to cause significant source-compatibility problems, but
there's at least some potential for it.  I think it's reasonable
to apply this regardless of whether the build mode is known to
require serialization; we really don't want import behavior to
change based on that kind of setting.