[5.2] SILGen: Fix withoutActuallyEscaping of 'c' closures
They don't have a context and therefore are not consumed.

Fixes a failing assert.

* Description: The compiler asserts on code that uses
withoutActuallyEscaping on a ‘c’ convention closure.

* Scope: Compiling source code that uses that combination will assert
and fail to compile. In no asserts build the compilation succeeds and
generates correct code (i.e the assert failure is inconsequential). The
problem existed in previously released version of the compiler.

* Testing: A swift regression test was added.

* Reviewed: Andrew T

Original PR: #29569

rdar://59046275